+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export * from './Desk';
+export * from './Effect';
+export * from './EffectParameter';
+export * from './Fader';
+export * from './Icon';
+export * from './Meter';
+export * from './Radio';
+export * from './Track';
